# docs/openapi.yaml
openapi: 3.0.3
info:
  title: My Social Networks API
  version: 1.0.0
  description: |
    API REST du nouveau service d'événements et de groupes de Facebook (TP Node.js / Express / MongoDB).

    **Authentification** : `POST /auth/register` ou `POST /auth/login` renvoient un token JWT (valable 24h)
    à envoyer dans le header `Authorization: Bearer <token>`.

    **Format des erreurs** : toutes les erreurs ont la forme `{ code, message, errors }`, où `errors`
    est la liste des problèmes à corriger (ex : erreurs de validation).

    **Listes** : les listes sont paginées avec `?page=1&limit=20` (100 maximum) et renvoient
    `{ data, pagination: { page, limit, total, pages } }`.

    **Visibilité** :
    - un groupe *secret* ou un événement *privé* répond `404` à ceux qui n'en font pas partie (il « n'existe pas » pour eux) ;
    - un groupe *privé* est visible (nom, description) mais ses membres, son fil et ses événements sont réservés aux membres (`403`).
  contact:
    name: Saad Lagzouli
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Développement
...
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [Auth]
      summary: Créer un compte
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterInput' }
      responses:
        '201':
          description: Compte créé, l'utilisateur est connecté
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Se connecter
      description: Limité à 20 tentatives par IP toutes les 15 minutes (`429` au-delà).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { description: Trop de tentatives }
  /auth/me:
    get:
      operationId: me
      tags: [Auth]
      summary: Utilisateur connecté
      responses:
        '200':
          description: Profil complet (avec email)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------------ Users
...
  /events/{id}/albums:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getAlbums
      tags: [Albums]
      summary: Albums de l'événement
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AlbumList' }
    post:
      operationId: createAlbum
      tags: [Albums]
      summary: Créer un album (participant)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlbumInput' }
      responses:
        '201':
          description: Album créé
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Album' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /albums/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getAlbumById
      tags: [Albums]
      summary: Détail d'un album
      responses:
        '200':
          description: Album avec `photos_count`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Album' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateAlbum
      tags: [Albums]
      summary: Modifier un album (créateur ou organisateur)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AlbumInput' }
      responses:
        '200': { description: Album modifié }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deleteAlbum
      tags: [Albums]
      summary: Supprimer un album, ses photos et leurs commentaires (créateur ou organisateur)
      responses:
        '200': { description: Album supprimé }
        '403': { $ref: '#/components/responses/Forbidden' }
  /albums/{id}/photos:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getPhotos
      tags: [Albums]
      summary: Photos de l'album
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PhotoList' }
    post:
      operationId: createPhoto
      tags: [Albums]
      summary: Poster une photo (participant)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PhotoInput' }
      responses:
        '201':
          description: Photo ajoutée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /photos/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getPhotoById
      tags: [Albums]
      summary: Détail d'une photo
      responses:
        '200':
          description: Photo avec `comments_count`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Photo' }
    patch:
      operationId: updatePhoto
      tags: [Albums]
      summary: Modifier la légende (auteur)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption: { type: string, maxLength: 500, nullable: true }
      responses:
        '200': { description: Photo modifiée }
        '403': { $ref: '#/components/responses/Forbidden' }
    delete:
      operationId: deletePhoto
      tags: [Albums]
      summary: Supprimer une photo (auteur ou organisateur)
      responses:
        '200': { description: Photo supprimée }
        '403': { $ref: '#/components/responses/Forbidden' }
  /photos/{id}/comments:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    get:
      operationId: getComments
      tags: [Albums]
      summary: Commentaires de la photo
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommentList' }
    post:
      operationId: createComment
      tags: [Albums]
      summary: Commenter une photo (participant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, maxLength: 1000, example: Superbe vue ! }
      responses:
        '201':
          description: Commentaire ajouté
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /photos/{id}/comments/{commentId}:
    delete:
      operationId: deleteComment
      tags: [Albums]
      summary: Supprimer un commentaire (auteur ou organisateur)
      parameters:
        - $ref: '#/components/parameters/Id'
        - { name: commentId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Commentaire supprimé }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ------------------------------------------------------------------ Polls
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, example: 6702f1c2a1b2c3d4e5f60718 }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: Données invalides
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 400, message: Bad Request, errors: ['end_date doit être après start_date'] }
    Unauthorized:
      description: Token absent, invalide ou expiré
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Action non autorisée pour cet utilisateur
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 403, message: Forbidden, errors: ["action réservée aux organisateurs de l'événement"] }
    NotFound:
      description: Ressource introuvable (ou invisible pour cet utilisateur)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 404, message: Event Not Found }
    Conflict:
      description: Conflit avec l'état actuel (doublon, plus de place, règle métier)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: 409, message: Conflict, errors: ["une personne ne peut obtenir qu'un seul billet par événement"] }
  schemas:
    Error:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        errors: { type: array, items: { type: string } }
    Pagination:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        pages: { type: integer, example: 3 }
    ObjectId:
      type: string
      example: 6702f1c2a1b2c3d4e5f60718
    Timestamps:
      type: object
      properties:
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ---------- users
    UserPublic:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        firstname: { type: string, example: Saad }
        lastname: { type: string, example: Lagzouli }
        avatar: { type: string, format: uri, nullable: true }
        bio: { type: string, nullable: true }
    User:
      allOf:
        - $ref: '#/components/schemas/UserPublic'
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            email: { type: string, format: email }
            birthdate: { type: string, format: date-time, nullable: true }
    RegisterInput:
      type: object
      required: [firstname, lastname, email, password]
      properties:
        firstname: { type: string, maxLength: 50, description: 'Lettres, espaces, tirets, apostrophes', example: Saad }
        lastname: { type: string, maxLength: 50, example: Lagzouli }
        email: { type: string, format: email, description: Unique, example: saad@exemple.fr }
        password: { type: string, minLength: 8, maxLength: 72, description: Au moins une lettre et un chiffre, example: motdepasse1 }
        avatar: { type: string, format: uri, nullable: true }
        bio: { type: string, maxLength: 500, nullable: true }
        birthdate: { type: string, format: date, nullable: true, description: Date passée }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: saad@exemple.fr }
        password: { type: string, example: motdepasse1 }
...
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: JWT à envoyer dans Authorization }
        user: { $ref: '#/components/schemas/User' }
...
    AlbumInput:
      type: object
      required: [name]
      properties:
        name: { type: string, maxLength: 100, example: Photos de la soirée }
        description: { type: string, maxLength: 1000, nullable: true }
    Album:
      allOf:
        - $ref: '#/components/schemas/AlbumInput'
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            event: { $ref: '#/components/schemas/ObjectId' }
            created_by: { $ref: '#/components/schemas/UserPublic' }
            photos_count: { type: integer }
    AlbumList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Album' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    PhotoInput:
      type: object
      required: [url]
      properties:
        url: { type: string, format: uri, example: 'https://exemple.com/photo.jpg' }
        caption: { type: string, maxLength: 500, nullable: true }
    Photo:
      allOf:
        - $ref: '#/components/schemas/PhotoInput'
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          properties:
            _id: { $ref: '#/components/schemas/ObjectId' }
            album: { $ref: '#/components/schemas/ObjectId' }
            event: { $ref: '#/components/schemas/ObjectId' }
            author: { $ref: '#/components/schemas/UserPublic' }
            comments_count: { type: integer }
    PhotoList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Photo' } }
        pagination: { $ref: '#/components/schemas/Pagination' }
    Comment:
      type: object
      properties:
        _id: { $ref: '#/components/schemas/ObjectId' }
        photo: { $ref: '#/components/schemas/ObjectId' }
        author: { $ref: '#/components/schemas/UserPublic' }
        content: { type: string }
        created_at: { type: string, format: date-time }
    CommentList:
      type: object
      properties:
        data: { type: array, items: { $ref: '#/components/schemas/Comment' } }
        pagination: { $ref: '#/components/schemas/Pagination' }

    # ---------- polls
...